+++ substrafl_examples/get_started/run_mnist_torch.py
@@ -474,14 +474,8 @@ compute_plan = execute_experiment(
 # Explore the results
 # *******************
 
-import time
-
-# if we are using remote clients, we have to wait until the compute plan is done before getting the results
-while (
-    client_0.get_compute_plan(compute_plan.key).status == "PLAN_STATUS_DOING"
-    or client_0.get_compute_plan(compute_plan.key).status == "PLAN_STATUS_TODO"
-):
-    time.sleep(2)
+# The results will be available once the compute plan is completed
+client_0.wait_compute_plan(compute_plan.key)
 # %%
 # List results
 # ============